Services/wms-monolith/pyproject.toml
[project]
name = "WMS"
version = "0.1.0"
description = "A WMS project for learning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "gunicorn>=23.0.0",
    "uvicorn>=0.40.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=5.0.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "email-validator>=2.3.0",
    "langchain>=0.1.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.0.20",
    "langchain-groq>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.2.0",
    "faker>=24.0.0",
    "langchain-huggingface>=0.0.1",
    "faiss-cpu>=1.7.4",
    "sqlparse>=0.4.4",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "langchain-text-splitters>=0.0.1",
    "sentence-transformers>=2.2.2",
    "tiktoken>=0.5.2",
    "transformers>=4.36.0",
    "pathlib2>=2.3.7",
    "black>=23.0.0",
    "mypy>=1.7.0",
    "tree-sitter>=0.25.2",
    "tree-sitter-languages>=1.10.2",
    "torch>=2.11.0",
    "bs4>=0.0.2",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.3",
    "rank-bm25>=0.2.2",
    "alembic>=1.18.4",
    "grpcio>=1.80.0",
    "grpcio-tools>=1.80.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1",
    "requests>=2.32.5",
    "aiosqlite>=0.20.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/app","src/ai_engine"]

[tool.hatch.build.targets.editable]
packages = ["src/app","src/ai_engine"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
